refactoring without signedOrder
