+++ src/api.rs
@@ -1,4 +1,3 @@
 pub mod api;
 pub mod filter;
 pub mod handler;
-pub use api::SignedOrder;
